Author monad refactor
